requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.0
decorator==5.1.1
Django==4.2.14
django-cors-headers==4.4.0
django-filter==24.2
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
idna==3.7
inflection==0.5.1
ipython==8.12.3
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
stripe==10.5.0
tomli==2.0.1
traitlets==5.14.3
types-PyYAML==6.0.12.20240311
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
